refactor: flatten app context providers
